fix #6

+++ database/migrations/2018_04_23_084602_create_results_table.php
@@ -17,8 +17,8 @@ class CreateResultsTable extends Migration
             $table->increments('id');
             $table->string('course');
             $table->integer('blok');
-            $table->integer('grade');
-            $table->integer('ec');
+            $table->decimal('grade');
+            $table->decimal('ec');
             $table->timestamps();
         });
     }
